Locations/events api structure
