+++ backend/src/mappening/api/locations.py
@@ -8,9 +8,11 @@ import re
 import json
 import os
 from operator import itemgetter
+import location_utils
+from mappening.utils import tokenize
 
 # Must be on same level as app.py
-data = json.load(open('sampleData.json'))
+# data = json.load(open('sampleData.json'))
 
 locations = Blueprint('locations', __name__)
 
@@ -33,113 +35,28 @@ def get_all_locations():
     # street, and zip for each location
     return jsonify({'locations': output})
 
-# Go through all events in events db and extract unique locations from the events
-def find_locations():
-    # Iterate through all events and get list of unique venues
-    places = []
-    place = {}
+# UPDATE DATABASE
 
-    # Latitude and Longitude range from (-90, 90) and (-180, 180)
-    INVALID_COORDINATE = 420
-
-    # Set parameters for Bruin Bear/Center of UCLA
-    CENTER_LATITUDE = "34.070966"
-    CENTER_LONGITUDE = "-118.445"
-    
-    # TODO: Integrate with new events caller
-    # Every time there are new events, check location info and update db if necessary
-    # events_cursor = events_ml_collection.find({"place": {"$exists": True}})
-    events_cursor = total_events_collection.find({"place": {"$exists": True}})
-    if events_cursor.count() > 0:
-      for event in events_cursor:
-        # Add location info to place dict
-        if 'location' in event['place']:
-          place['location'] = event['place']['location']
-        if 'name' in event['place']:
-          place['location']['name'] = event['place']['name']
-
-        # Check that place is not empty
-        if any(place):
-          # All places should have alternative names field
-          place['location']['alternative_names'] = []
-          if 'name' in event['place']:
-            place['location']['alternative_names'].append(place['location']['name'])
-          # Reject exact matches
-          if place not in places:
-            # Check whether coordinates exist
-            if 'latitude' in place['location'] and 'longitude' in place['location']:
-              # Check whether coordinates match another event
-              if not any(loc.get('latitude', INVALID_COORDINATE) == place['location']['latitude'] and loc.get('longitude', INVALID_COORDINATE) == place['location']['longitude'] for loc in places):
-                # No matching coordinates, append to list
-                places.append(place)
-              else:
-                # There exists a nonidentical location with matching coordinates
-                # Merge information of new event into old event
-                # Can probably improve this part
-                for loc in places:
-                  if loc['location'].get('latitude', INVALID_COORDINATE) == place['latitude'] and loc['location'].get('longitude', INVALID_COORDINATE) == place['longitude']:
-                    # Go through all the keys
-                    for key in place['location']:
-                      # If new key then add it to location
-                      if key not in loc['location']:
-                        loc['location'][key] = place['location'][key]
-                      # If names do not match, coordinates do so add name as an alternate name
-                      if key == "name" and 'name' in loc['location'] and loc['location']['name'] != place['location']['name']:
-                        if place['location']['name'].lower() not in (name.lower() for name in loc['location']['alternative_names']):
-                          if place['location']['name']:
-                            loc['location']['alternative_names'].append(place['location']['name'])
-            else:
-              # No coordinates exist, indicate that coordinates may be incorrect
-              place['coordinates'] = "GOOGLE"
-              # Try to get coordinates from google places
-              if 'name' in place['location']:
-                # Use location name to try to find location info
-                search_results = google_textSearch(place['location']['name'])
-                if search_results:
-                  # If there are results see if there is a latitude/longitude
-                  if search_results[0]['latitude'] == "NO LATITUDE" or search_results[0]['longitude'] == "NO LONGITUDE":
-                    place['location']['latitude'] = CENTER_LATITUDE
-                    place['location']['longitude'] = CENTER_LONGITUDE
-                  else:
-                    place['location']['latitude'] = search_results[0]['latitude']
-                    place['location']['longitude'] = search_results[0]['longitude']
-              # If there is no name, see if there is street info
-              elif 'street' in place['location'] and place['location']['street'] != "NO STREET" and place['location']['street'] != '':
-                # Use name to try to find location info
-                search_results = google_textSearch(place['location']['street'])
-                if search_results:
-                  if search_results[0]['latitude'] == "NO LATITUDE" or search_results[0]['longitude'] == "NO LONGITUDE":
-                    place['location']['latitude'] = CENTER_LATITUDE
-                    place['location']['longitude'] = CENTER_LONGITUDE
-                  else:
-                    place['location']['latitude'] = search_results[0]['latitude']
-                    place['location']['longitude'] = search_results[0]['longitude']
-                    updated = True
-              else:
-                # There was no name or street info, set to Bruin Bear location
-                place['location']['latitude'] = CENTER_LATITUDE
-                place['location']['longitude'] = CENTER_LONGITUDE
-
-              places.append(place)
-        # Reset place to an empty dict
-        place = {}
-      return places
-    else:
-      return 'Cannot find any events with locations!'
-
-# Add locations to mlab db
-# Gives some duplicate events in db
-# Duplicate events resolved if the different names used are put under alternate_names
-@locations.route('/db_locations')
-def db_locations():
+# Add locations to database from given collection
+# Sample collection(s): events_ml_collection, ucla_events_collection
+# TODO: hook up so everytime we get new events we add their location data to db
+@locations.route('/add/<events_collection>', methods=['PUT'])
+def add_locations_from_collection(events_collection):
     # Update locations or insert new locations from events in db
-    new_locations = find_locations()
     updated_locations = []
     added_locations = []
     updated = False
 
+    # TODO  Verify collection is valid
+    if events_collection != "ucla_events" and events_collection != "events_ml" and events_collection != "test":
+      return jsonify({'Added Locations': added_locations, 'Updated Locations': updated_locations})
+
+    new_locations = location_utils.get_locations_from_collection(events_collection)
+
     # Latitude and Longitude range from (-90, 90) and (-180, 180)
     INVALID_COORDINATE = 420
+
+    print new_locations
     
     # For every location from events db
     for new_loc in new_locations:
@@ -147,7 +64,7 @@ def db_locations():
       place_name = re.sub(r'\bUCLA-\s?', '', new_loc['location'].get('name', "NO NAME"), flags=re.IGNORECASE)
       place_name = re.sub(r'-UCLA\s?', '', place_name, flags=re.IGNORECASE)
       place_name = re.sub(r'\b[a-zA-Z]+\d+\s?', '', place_name, flags=re.IGNORECASE)
-      place_name = process.processText(place_name)
+      place_name = tokenize.tokenize_text(place_name)
       processed_place = re.compile(place_name, re.IGNORECASE)
 
       # Find location of same coordinates/name
@@ -203,440 +120,40 @@ def db_locations():
 
     return jsonify({'Added Locations': added_locations, 'Updated Locations': updated_locations})
 
-# Given a location string try to return coordinates/relevant location info
-# e.g. BH 3400 => Boelter Hall vs. Engr 4 => Engineering IV vs. Engineering VI
-@locations.route('/coordinates/<place_query>', methods=['GET'])
-def get_mongo_textSearch(place_query):
-    output = []
-    output_places = []
-
-    # Supplied string such as "Boelter Hall" for a location
-    print "Original place query: " + place_query
-    # Remove leading/trailing white space
-    place_query = place_query.strip()
-
-    # Search for exact match first
-    # Sometimes regency village weighted more than sunset village due to repetition of village
-    # place_regex = re.compile('.*' + place_query + '.*', re.IGNORECASE)
-    place_regex = place_query
-    if place_query.lower() != "ucla":
-      place_regex = re.sub(r'\bUCLA\s?', '', place_regex, flags=re.IGNORECASE)
-    place_regex = re.sub(r'\|', ' ', place_regex, flags=re.IGNORECASE)
-    place_regex = re.sub(r'\(', '', place_regex, flags=re.IGNORECASE)
-    place_regex = re.sub(r'\)', '', place_regex, flags=re.IGNORECASE)
-    place_regex = place_regex.strip()
-    print "Regex place query: " + place_regex
-
-    place_regex = re.compile("^" + place_regex + "$", re.IGNORECASE)
-    places_cursor = UCLA_locations_collection.find({'location.alternative_names': place_regex})
-    
-    # Places that match the name are appended to output
-    if places_cursor.count() > 0:
-      for place in places_cursor:
-        output.append({
-          'name': place['location'].get('name', "NO NAME"),
-          'street': place['location'].get('street', "NO STREET"),
-          'zip': place['location'].get('zip', "NO ZIP"),
-          'city': place['location'].get('city', "NO CITY"),
-          'state': place['location'].get('state', "NO STATE"),
-          'country': place['location'].get('country', "NO COUNTRY"),
-          'latitude': place['location'].get('latitude', "NO LATITUDE"),
-          'longitude': place['location'].get('longitude', "NO LONGITUDE"),
-          'alternative_names': place['location']['alternative_names']
-        })
-        output_places.append(place['location'].get('name', "NO NAME"))
-
-      print "Found exact match!"
-      return output
-
-    print "Doing text search..."
-
-    # Tokenize and remove unnecessary/common words 
-    place_name = re.sub(r'\bUCLA-\s?', '', place_query, flags=re.IGNORECASE)
-    place_name = re.sub(r'-UCLA\s?', '', place_name, flags=re.IGNORECASE)
-    place_name = re.sub(r'\b[a-zA-Z]+\d+\s?', '', place_name, flags=re.IGNORECASE)
-    processed_place = process.processText(place_name)
-    print "Processed place query: " + processed_place
-
-    # Locations db has text search index on alternate_locations field
-    # Search for locations that match words in processed place query
-    # Default stop words for english language, case insensitive
-    # Sort by score (based on number of occurances of query words in alternate names)
-    # Can limit numer of results as well
-    places_cursor = UCLA_locations_collection.find( 
-      { '$text': { '$search': processed_place, '$language': 'english', '$caseSensitive': False } },
-      { 'score': { '$meta': 'textScore' } }
-    ).sort([('score', { '$meta': 'textScore' })]) #.limit(3)
-
-    # Places that match the alternate name are appended to output if not already
-    # part of output
-    if places_cursor.count() > 0:
-      for place in places_cursor:
-        # Check if already added by maintaining list of places added by name
-        if place['location'].get('name', "NO NAME") not in output_places:
-          output.append({
-            'score': place['score'],
-            'name': place['location'].get('name', "NO NAME"),
-            'street': place['location'].get('street', "NO STREET"),
-            'zip': place['location'].get('zip', "NO ZIP"),
-            'city': place['location'].get('city', "NO CITY"),
-            'state': place['location'].get('state', "NO STATE"),
-            'country': place['location'].get('country', "NO COUNTRY"),
-            'latitude': place['location'].get('latitude', "NO LATITUDE"),
-            'longitude': place['location'].get('longitude', "NO LONGITUDE"),
-            'alternative_names': place['location']['alternative_names']
-          })
-          output_places.append(place['location'].get('name', "NO NAME"))
-
-    return output
-
-# Pretty output of location search results
-@locations.route('/location_search/<place_query>', methods=['GET'])
-def get_location_results(place_query):
-    output = get_mongo_textSearch(place_query)
-    return jsonify({"Database Results": output}) 
+# LOCATIONS SEARCH
 
 # Given location name, return location data or some indication that no location
-# could be found. use on tkinter_UCLA_locations_collection for testing/metrics.
-@locations.route('/location_search_result/<place_query>', methods=['GET'])
-def get_location_search_result(place_query):
-    output = get_mongo_textSearch(place_query)
+# could be found. Returns <num_results> top results
+@locations.route('/search/<place_query>', defaults={'num_results': None}, methods=['GET'])
+@locations.route('/search/<place_query>/<int:num_results>', methods=['GET'])
+def get_location_results(place_query, num_results):
+    search_results = location_utils.search_locations(place_query)
 
-    if not output:
+    if not search_results:
       return "There were no results!"
+    elif not num_results or num_results <= 0:
+      return jsonify({"Locations": search_results}) 
     else:
-      return output[0]
+      output = []
+      for i in range(0, num_results):
+        output.append(search_results[i])
+      return jsonify({'Locations': output})
 
-# Go through tkinter_UCLA_locations_collection and search every location name. Manually
-# verify correctness or just look at cases that don't match
-# TODO: process tkinter_UCLA_locations with tkinterUCLA.py before adding to tkinter_unknown_locations
-# curl -d -X POST http://localhost:5000/test_unknown_locations
-@locations.route('/test_unknown_locations', methods=['POST'])
-def test_unknown_locations():
-  num_assigned = 0
-  num_unassigned = 0
-  counter = 1
-
-  locs_cursor = tkinter_UCLA_locations_collection.find({}, {'_id': False})
-  if locs_cursor.count() > 0:
-    for loc_db in locs_cursor:
-      print "~~~~~~~ " + str(counter) + " ~~~~~~~" + " WR: " + str(num_unassigned)
-      counter = counter + 1
-      if 'location_name' in loc_db:
-        loc_result = get_location_search_result(loc_db['location_name'])
-        if loc_result != "There were no results!":
-          print "Found a match!"
-          num_assigned = num_assigned + 1
-          tkinter_unknown_locations_collection.insert_one({
-            "unknown_loc": {
-              "loc_name": loc_db.get('location_name', "NO LOCATION NAME"),
-              "event_name": loc_db.get('event_name', "NO EVENT NAME")
-            },
-            "db_loc": {
-              "loc_name": loc_result['name'],
-              "loc_alt_names": loc_result['alternative_names'],
-              "loc_latitude": loc_result['latitude'],
-              "loc_longitude": loc_result['longitude'],
-              "map_url": "https://www.google.com/maps/place/" + str(loc_result['latitude']) + "," + str(loc_result['longitude'])
-            }
-          }) 
-        else:
-          print "Didn't find a location!"
-          num_unassigned = num_unassigned + 1
-          tkinter_TODO_locations_collection.insert_one({
-            "unknown_loc": {
-              "loc_name": loc_db.get('location_name', "NO LOCATION NAME"),
-              "event_name": loc_db.get('event_name', "NO EVENT NAME")
-            },
-            "db_loc": {}
-          }) 
-  else:
-      print 'Cannot find any unknown locations!'
-
-  print "num_assigned: " + num_assigned
-  print "num_unassigned: " + num_unassigned
-  return "Added unknown locations to database\n"
-
-# Go through events_ml_collection and search every location name. See if result
-# matches what we expected for metric purposes.
-@locations.route('/test_locations', methods=['GET'])
-def test_locations_api():
-  num_correct = 0
-  num_wrong = 0
-  num_invalid = 0
-  wrong_locs = []
-  counter = 1
-
-  events_cursor = events_ml_collection.find({}, {'_id': False})
-  if events_cursor.count() > 0:
-    for event in events_cursor:
-      print "~~~~~~~ " + str(counter) + " ~~~~~~~" + " WR: " + str(num_wrong)
-      counter = counter + 1
-      if 'place' in event and 'location' in event['place']:
-        loc = get_location_search_result(event['place']['name'])
-        if loc != "There were no results!":
-          # Latitude and longitude are significant to the 4th digit, but 3rd to be safe
-          event_lat = event['place']['location'].get('latitude', 420)
-          event_long = event['place']['location'].get('longitude', 420)
-          lat_diff = abs(loc['latitude'] - event_lat)
-          long_diff = abs(loc['longitude'] - event_long)
-          if lat_diff < 0.0015 and long_diff < 0.0015:
-            print "Correct"
-            num_correct = num_correct + 1
-          else:
-            print "Wrong"
-            num_wrong = num_wrong + 1
-            wrong_locs.append({
-              "event": {
-                "place_name": event['place']['name'], 
-                "place_latitude": event['place']['location'].get('latitude', 420),
-                "place_longitude": event['place']['location'].get('longitude', 420)
-              },
-              "loc": {
-                "loc_name": loc['name'],
-                "loc_alt_names": loc['alternative_names'],
-                "loc_latitude": loc['latitude'],
-                "loc_longitude": loc['longitude']
-              }
-            }) 
-        else:
-          print "Invalid Loc"
-          num_invalid = num_invalid + 1
-          wrong_locs.append({
-              "event": {
-                "place_name": event['place']['name'], 
-                "place_latitude": event['place']['location'].get('latitude', 420),
-                "place_longitude": event['place']['location'].get('longitude', 420)
-              },
-              "loc": "NONE"
-            }) 
-  else:
-      print 'Cannot find any events!'
-
-  # Output typically contains name, city, country, latitude, longitude, state, 
-  # street, and zip for each location
-  return jsonify({'num_correct': num_correct, 'num_wrong': num_wrong, 'num_invalid': num_invalid, 'wrong_locs': wrong_locs})
-
-# Insert locations from a JSON file to the db
-# See sample format in ./sampleData.json
-# curl -d -X POST http://localhost:5000/insert_locations
-# TODO make other things that should be POST post lmao
-@locations.route('/insert_locations', methods=['POST'])
-def insert_locations():
-  UCLA_locations_collection.insert_many(data['locations'])
-  return "Successfully inserted location documents to db!"
-
-
-# Add tokenized version of all alternate names to alternate names list
-@locations.route('/tokenize_names', methods=['GET'])
-def get_tokenized_names():
-  places = []
-  updated = False
-
-  # Go through every location in json
-  for location in data['locations']:
-    place = location
-    if 'alternative_names' in place['location']:
-      for alt_name in place['location']['alternative_names']:
-        processed_name = process.processText(alt_name)
-        if processed_name not in (name.lower() for name in place['location']['alternative_names']):
-          if processed_name:
-            place['location']['alternative_names'].append(processed_name)
-            updated = True
-      if updated:
-        places.append(place)
-        updated = False
-
-  return jsonify({"locations": places})
-
-# Add tokenized version of all alternate names to alternate names list
-@locations.route('/UCLA_names', methods=['GET'])
-def get_UCLA_names():
-  places = []
-  loc_counter = 1
-
-  # Go through every location in json
-  for location in data['locations']:
-    loc_counter = loc_counter + 1
-    place = location
-    if 'alternative_names' in place['location']:
-      for alt_name in place['location']['alternative_names']:
-        processed_name = re.sub(r'\bUCLA\s?', '', alt_name, flags=re.IGNORECASE)
-        processed_name = processed_name.strip()
-        if processed_name.lower() not in (name.lower() for name in place['location']['alternative_names']):
-          if processed_name:
-            place['location']['alternative_names'].append(processed_name)
-      places.append(place)
-
-  return jsonify({"locations": places})
-
-# Fill out json data using google api results
-# Supplied json missing street/zip/latitude/longitude
-# Tries to fill out those missing fields using name field or street if supplied
-# Takes top result to fill out info, MAY BE INCORRECT
-# Warning: pretty slow, limit of 1000 requests? 100 requests?
-@locations.route('/location_data', methods=['GET'])
-def get_location_data():
-  places = []
-  updated_places = []
-  updated = False
-
-  # Go through every location in json
-  for location in data['locations']:
-    place = location
-    # Add stripped down name to alternative_names
-    if 'name' in location['location']:
-      processed_place = process.processText(location['location']['name'])
-      if 'alternative_names' in location['location']:
-        if location['location']['name'].lower() not in (name.lower() for name in location['location']['alternative_names']):
-          if location['location']['name']:
-            place['location']['alternative_names'].append(location['location']['name'])
-            updated = True
-        if processed_place not in (name.lower() for name in location['location']['alternative_names']):
-          if processed_place:
-            place['location']['alternative_names'].append(processed_place)
-            updated = True
-    # No street or zip information, try to find it
-    if 'street' not in location['location'] or 'zip' not in location['location'] or location['location']['street'] == '' or location['location']['zip'] == '':
-      if 'name' in location['location']:
-        # Use location name to try to find location info
-        search_results = google_textSearch(location['location']['name'])
-        if search_results:
-          # Assume first result is best result/most relevant result
-          # Set street to the address
-          place['location']['street'] = search_results[0]['address']
-
-          # Extract zip code from address
-          re_result = re.search(r'(\d{5}(\-\d{4})?)', place['location']['street'])
-          if re_result:
-            place['location']['zip'] = re_result.group(0) # Sometimes get 5 digit address numbers
-          else:
-            place['location']['zip'] = "NO ZIP"
-          updated = True
-      else:
-        # Without a name, street, or zip cannot find out much about location
-        # Is it even a location at this point lmao
-        place['location']['street'] = "NO STREET"
-        place['location']['zip'] = "NO ZIP"
-        place['location']['name'] = "NO NAME"
-    # Check if latitude/longitude is filled out (420 is default value)
-    if 'latitude' not in location['location'] or 'longitude' not in location['location'] or location['location']['latitude'] == 420 or location['location']['longitude'] == 420:
-      if 'name' in location['location']:
-        # Use location name to try to find location info
-        search_results = google_textSearch(location['location']['name'])
-        if search_results:
-          # If there are results see if there is a latitude/longitude
-          if search_results[0]['latitude'] == "NO LATITUDE" or search_results[0]['longitude'] == "NO LONGITUDE":
-            place['location']['latitude'] = 404
-            place['location']['longitude'] = 404
-          else:
-            place['location']['latitude'] = search_results[0]['latitude']
-            place['location']['longitude'] = search_results[0]['longitude']
-            updated = True
-      # If there is no name, see if there is street info
-      elif 'street' in place['location'] and place['location']['street'] != "NO STREET" and place['location']['street'] != '':
-        # Use name to try to find location info
-        search_results = google_textSearch(place['location']['street'])
-        if search_results:
-          if search_results[0]['latitude'] == "NO LATITUDE" or search_results[0]['longitude'] == "NO LONGITUDE":
-            place['location']['latitude'] = 404
-            place['location']['longitude'] = 404
-          else:
-            place['location']['latitude'] = search_results[0]['latitude']
-            place['location']['longitude'] = search_results[0]['longitude']
-            updated = True
-      else:
-        # There was no name or street info, set to another junk value
-        place['location']['latitude'] = 666
-        place['location']['longitude'] = 666
-
-    # If we want to keep track of all places from json data uncomment this
-    # places.append(place)
-
-    # Keep track of places with info that was actually updated
-    if updated:
-      updated_places.append(place)
-    updated = False
-
-  # Return json info on updated locations and/or all locations from json data
-  # return jsonify({"locations": places, "changed locations": updated_places})
-  return jsonify({"New/Modified Locations": updated_places})
-
-# Use the Google Places API Web Service to get info about places based on a string
-# Results ordered by perceived relevance (use UCLA in query?)
-# Can bias results by specifying location (latitude/longitude) and radius (meters)
-def google_textSearch(place_query):
-    # Set parameters for Bruin Bear/Center of UCLA
-    CENTER_LATITUDE = "34.070966"
-    CENTER_LONGITUDE = "-118.445"
-
-    # For comprehension: School of Theater, Film, TV within radius 700
-    # Hammer Museum within radius 1300, Saffron and Rose within radius 1800
-    RADIUS = "2000"
-
-    # Base URL
-    TextSearch_URL = 'https://maps.googleapis.com/maps/api/place/textsearch/json?'
-
-    # Forming the request
-    textSearch = TextSearch_URL + "query=" + place_query + "&location=" + CENTER_LATITUDE + "," + CENTER_LONGITUDE + "&radius=" + RADIUS + "&key=" + GOOGLE_API_KEY
-
-    # Getting result from URL and processing
-    resultsPage = requests.get(textSearch)
-    resultsJSON = json.loads(resultsPage.content)
-
-    output = []
-    for result in resultsJSON['results']:
-      # Extract most relevant information from JSON results
-      output.append({
-          'name': result.get('name', "NO NAME"),
-          'address': result.get('formatted_address', "NO ADDRESS"),
-          'latitude': result['geometry']['location'].get('lat', "NO LATITUDE"),
-          'longitude': result['geometry']['location'].get('lng', "NO LONGITUDE")
-      })
-
-    return output
+# GOOGLE WRAPPER 
 
 # Run Google Maps TextSearch on given query and print all results in JSON
-# Uses same function as before but for printing rather than use in event processing
-@locations.route('/place_textSearch/<place_query>', methods=['GET'])
+# Print all results in JSON, a wrapper for Google's API
+@locations.route('/google/search/text/<place_query>', methods=['GET'])
 def get_textsearch(place_query):
-    output = google_textSearch(place_query)
+    output = location_utils.google_textSearch(place_query)
 
     return jsonify({'results': output})
-    # return jsonify({'results': resultsJSON})
 
 # Run Google Maps NearbySearch on given query. Uses Google Places API Web Service
-# to search for places within specified area. Takes in location, radius, and a keyword.
-# Optional rankby parameter, and also returns more results than textSearch
-@locations.route('/place_nearbySearch/<place_query>', methods=['GET'])
+# to search for places within specified area. Returns more results than textSearch
+# Print all results in JSON, a wrapper for Google's API
+@locations.route('/google/search/nearby/<place_query>', methods=['GET'])
 def get_nearbysearch(place_query):
-    # Set parameters for Bruin Bear/Center of UCLA
-    CENTER_LATITUDE = "34.070966"
-    CENTER_LONGITUDE = "-118.445"
-
-    # For comprehension: School of Theater, Film, TV within radius 700
-    # Hammer Museum within radius 1300, Saffron and Rose within radius 1800
-    RADIUS = "2000"
-
-    # Base URL
-    NearbySearch_URL = 'https://maps.googleapis.com/maps/api/place/nearbysearch/json?'
-
-    # Forming the request
-    nearbySearch = NearbySearch_URL + "location=" + CENTER_LATITUDE + "," + CENTER_LONGITUDE + "&radius=" + RADIUS + "&keyword=" + place_query + "&key=" + GOOGLE_API_KEY
-
-    # Getting result from URL and processing
-    resultsPage = requests.get(nearbySearch)
-    resultsJSON = json.loads(resultsPage.content)
-
-    output = []
-    for result in resultsJSON['results']:
-      # Extract most relevant information from JSON results
-      output.append({
-          'name': result.get('name', "NO NAME"),
-          'address': result.get('vicinity', "NO ADDRESS"),
-          'latitude': result['geometry']['location'].get('lat', "NO LATITUDE"),
-          'longitude': result['geometry']['location'].get('lng', "NO LONGITUDE")
-      })
+    output = location_utils.google_nearbySearch(place_query)
 
     return jsonify({'results': output})
-    # return jsonify({'results': resultsJSON})
